relocate settings page queries
